Secrets: move mocks from testutils file
