+++ pkg/registry/apis/secret/testutils/noop_migration_executor.go
@@ -1,0 +1,10 @@
+package testutils
+
+import "context"
+
+type NoopMigrationExecutor struct {
+}
+
+func (e *NoopMigrationExecutor) Execute(ctx context.Context) (int, error) {
+	return 0, nil
+}
